Z for RGB capture pausing.
This now needs the actual RGB frame to be 3D wherever there is depth (i.e., the moving person should extrude from the background), with some 3D optical flow (use the depth as a heightmap for lerping and checking Z-flow).

Lastly... we need that background RGB stuff to make it actually look right.